requirements.txt
colorama
progressbar
python-magic
